Renamed the final filenames for this time: RICursors		RiuCursors RIMainWindow	RiuMainWindow RIPreferencesDialog	RiuPreferencesDialog RIProcessMonitor	RiuProcessMonitor RIResultInfoPanel	RiuResultInfoPanel RIViewer		RiuViewer p4#: 21072
