+++ ApplicationCode/ProjectDataModel/RimUiTreeView.cpp
@@ -25,10 +25,10 @@
 #include "RimReservoirView.h"
 #include "RimCalcScript.h"
 #include "RiaApplication.h"
-#include "RIMainWindow.h"
+#include "RiuMainWindow.h"
 #include "RimInputPropertyCollection.h"
 #include "RimExportInputPropertySettings.h"
-#include "RIPreferencesDialog.h"
+#include "RiuPreferencesDialog.h"
 #include "RifEclipseInputFileTools.h"
 #include "RimInputCase.h"
 #include "RimBinaryExportSettings.h"
