+++ ApplicationCode/UserInterface/RiuResultInfoPanel.h
@@ -27,7 +27,7 @@ class QTextEdit;
 
 //==================================================================================================
 //
-// RIResultInfoPanel 
+// RiuResultInfoPanel 
 //
 //==================================================================================================
 class RiuResultInfoPanel : public QWidget
