+++ ApplicationCode/UserInterface/RiuViewer.h
@@ -38,7 +38,7 @@ namespace cvf
 
 //==================================================================================================
 //
-// RIViewer
+// RiuViewer
 //
 //==================================================================================================
 class RiuViewer : public caf::Viewer
